MSVC: cleanup Warning level 4

+++ Path.cpp
@@ -12,10 +12,10 @@ namespace Path
 {
 	const std::string Filename(const std::string& path, const bool getExt, const char delim)
 	{
-		int start = 0;
-		int end = (path.length() - 1);
+		size_t start = 0;
+		size_t end = (path.length() - 1);
 
-		for (int i = end; i >= 0; i--)
+		for (size_t i = end; i >= 0; i--)
 		{
 			if (path[i] == delim)
 			{
@@ -26,7 +26,7 @@ namespace Path
 
 		if (!getExt)
 		{
-			for (int i = start; i <= end; i++)
+			for (size_t i = start; i <= end; i++)
 			{
 				if (path[i] == '.')
 				{
@@ -41,10 +41,10 @@ namespace Path
 
 	const std::string Extension(const std::string& path, const bool getDot, const char delim)
 	{
-		int start = 0;
-		int end = (path.length() - 1);
+		size_t start = 0;
+		size_t end = (path.length() - 1);
 
-		for (int i = end; i >= 0; i--)
+		for (size_t i = end; i >= 0; i--)
 		{
 			if (path[i] == delim)
 			{
@@ -59,10 +59,10 @@ namespace Path
 	const std::string Directory(const std::string& path, const uint32_t cutDir, const bool addSuffix, const char delim)
 	{
 		int start = 0;
-		int end = (path.length() - 1);
+		size_t end = (path.length() - 1);
 		uint32_t cut = 0;
 
-		for (int i = end; i >= 0; i--)
+		for (size_t i = end; i >= 0; i--)
 		{
 			if (path[i] == delim)
 			{
